+++ saleor/graphql/payment/mutations/payment/checkout_payment_create.py
@@ -28,11 +28,11 @@ from ....core.mutations import BaseMutation
 from ....core.scalars import UUID, PositiveDecimal
 from ....core.types import BaseInputObjectType
 from ....core.types import common as common_types
+from ....core.utils.metadata_manager import MetadataManager
 from ....meta.inputs import MetadataInput
 from ....plugins.dataloaders import get_plugin_manager_promise
 from ...enums import StorePaymentMethodEnum
 from ...types import Payment
-from ...utils import metadata_contains_empty_key
 
 
 class PaymentInput(BaseInputObjectType):
@@ -177,7 +177,7 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
 
     @classmethod
     def validate_metadata_keys(cls, metadata_list: list[dict]):
-        if metadata_contains_empty_key(metadata_list):
+        if MetadataManager.metadata_contains_empty_key(metadata_list):
             raise ValidationError(
                 {
                     "input": ValidationError(
