+++ saleor/graphql/payment/utils.py
@@ -7,10 +7,6 @@ if TYPE_CHECKING:
     from ...app.models import App
 
 
-def metadata_contains_empty_key(metadata_list: list[dict]) -> bool:
-    return not all(data["key"].strip() for data in metadata_list)
-
-
 def check_if_requestor_has_access(
     transaction: payment_models.TransactionItem,
     user: Optional["User"],
